requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
httpx>=0.28
pydantic>=2.0
Pillow>=11.0
